orion-services/orion-calendar-importer/requirements.txt
Flask==3.0.0
google-cloud-firestore==2.16.0
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-secret-manager==2.19.0
gunicorn==22.0.0
requests==2.32.3
